add more overhead to nerf the bot

+++ discord/src/commands/image.rs
@@ -19,16 +19,16 @@ struct Image;
 /// Get the last image buf from most recent message on the channel
 /// Max messages length is 100
 #[inline]
-pub fn get_last_image_buf(ctx: &Context, msg: &Message, limit: u64) -> Option<Bytes> {
+pub fn get_last_image_buf(ctx: &Context, msg: &Message, limit: u16) -> Option<Bytes> {
     get_last_image_url(ctx, msg, limit).and_then(|v| get_file_bytes(v).ok())
 }
 
 #[inline]
-pub fn get_last_image_url(ctx: &Context, msg: &Message, limit: u64) -> Option<String> {
+pub fn get_last_image_url(ctx: &Context, msg: &Message, limit: u16) -> Option<String> {
     get_image_url_from_message(&msg).or_else(|| {
         let current_id = msg.id;
         msg.channel_id
-            .messages(ctx, |m| m.limit(limit).before(current_id))
+            .messages(ctx, |m| m.limit(limit as u64).before(current_id))
             .ok()
             .and_then(|msgs| {
                 msgs.into_iter()
